🧹 chore: separate otp generation as utility function
